refactor: relocate forwarding implementation
